[Kernel] Refactoring everything
Changing formatting to the same format & little optimizations

+++ Kernel/include/Kernel/Common/Types.hpp
@@ -12,6 +12,6 @@ namespace Kernel
 {
 	using ssize_t = int64_t;
 	
-	using pid_t = int64_t;
-	using tid_t = int64_t;
+	using pid_t = ssize_t;
+	using tid_t = ssize_t;
 }
